use existing macro for metadata reading

+++ sherpa-onnx/csrc/offline-canary-model-meta-data.h
@@ -15,7 +15,7 @@ struct OfflineCanaryModelMetaData {
   int32_t subsampling_factor = 8;
   int32_t feat_dim = 120;
   int32_t num_decoder_layers = 6;
-  int64_t decoder_hidden_size = 1024;
+  int32_t decoder_hidden_size = 1024;
   std::string normalize_type;
   std::unordered_map<std::string, int32_t> lang2id;
 };
